requirements.txt
streamlit==1.28.1
plotly==5.17.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
pymongo==4.3.3
stripe==5.4.0
twilio==7.16.4
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
cryptography==40.0.1
bcrypt==4.0.1